Issue #19: Restored prisma intitialization. Deploy tests.
